# app/backend/docs/swagger.yaml
basePath: /
definitions:
...
  handlers.ApiKeyRequest:
    properties:
      name:
        type: string
    required:
    - name
    type: object
  handlers.ApiKeyWithValueResponse:
    properties:
      created_at:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      key:
        type: string
      key_preview:
        type: string
      last_used:
        type: string
      name:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Ticket Management System API server
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: TMS API
  version: "1.0"
paths:
...
  /v1/tenants/{tenant_id}/projects/{project_id}/api-keys:
    get:
      consumes:
      - application/json
      description: Get a list of API keys for the project
      parameters:
      - description: Tenant ID
        format: uuid
        in: path
        name: tenant_id
        required: true
        type: string
      - description: Project ID
        format: uuid
        in: path
        name: project_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of API keys
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized - Invalid or missing authentication
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden - Insufficient permissions
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List API keys
      tags:
      - API Keys
    post:
      consumes:
      - application/json
      description: Create a new API key for the project
      parameters:
      - description: Tenant ID
        format: uuid
        in: path
        name: tenant_id
        required: true
        type: string
      - description: Project ID
        format: uuid
        in: path
        name: project_id
        required: true
        type: string
      - description: API key creation data
        in: body
        name: apikey
        required: true
        schema:
          $ref: '#/definitions/handlers.ApiKeyRequest'
      produces:
      - application/json
      responses:
        "201":
          description: API key created successfully (key value shown only once)
          schema:
            $ref: '#/definitions/handlers.ApiKeyWithValueResponse'
        "400":
          description: Bad request - Invalid input data
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized - Invalid or missing authentication
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden - Insufficient permissions
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create API key
      tags:
      - API Keys
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"